feat(backend): extract TLS to C
TLS is fundamentally OS-specific and will probably not stabilize outside std.
We need to forbid floating-point operations in the backend, since we would rather not save the floating-point registers on each call to `mcount`.
Unfortunately, Rust will not allow compiling existing targets with soft-float, since that generally changes the ABI, even though it would be fine specifically if we also built std with soft-float and didn't use floats in our API.
Alas, we have to resort to C for TLS.